pyproject.toml
[project]
name = "tempemail"
version = "0.1.0"
description = "gerador de email temporários"
authors = [
    {name = "paulindavzl",email = "paulochz01@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtpd (>=1.4.6,<2.0.0)",
    "jsonschema (>=4.24.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]


[tool.poetry.scripts]
tests = "run_tests:run"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.0.0"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
